gym_management_system_v4/backend/requirements.txt
Django==4.2
djangorestframework
django-cors-headers
python-decouple
requests
bcrypt
async-timeout

